Split hawk authentication utils out from activity stream.
This is in preparation for adding the wins api endpoint for data workspace.
