+++ activity_stream/tests/test_views.py
@@ -9,6 +9,8 @@ from rest_framework import status
 from rest_framework.reverse import reverse
 from rest_framework.test import APIClient
 
+from test_helpers.hawk_utils import hawk_auth_sender
+
 
 @pytest.fixture
 def api_client():
@@ -37,22 +39,6 @@ def _url_incorrect_path():
     return 'http://testserver' + reverse('activity-stream') + 'incorrect/'
 
 
-def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
-                 method='GET', content='', content_type=''):
-    credentials = {
-        'id': key_id,
-        'key': secret_key,
-        'algorithm': 'sha256',
-    }
-    return mohawk.Sender(
-        credentials,
-        url(),
-        method,
-        content=content,
-        content_type=content_type,
-    )
-
-
 @pytest.mark.parametrize(
     'get_kwargs,expected_json',
     (
@@ -60,7 +46,7 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If no X-Forwarded-For header
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender().request_header,
+                HTTP_AUTHORIZATION=hawk_auth_sender(_url()).request_header,
             ),
             {'detail': 'Incorrect authentication credentials.'},
         ),
@@ -68,7 +54,7 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If second-to-last X-Forwarded-For header isn't whitelisted
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender().request_header,
+                HTTP_AUTHORIZATION=hawk_auth_sender(_url()).request_header,
                 HTTP_X_FORWARDED_FOR='9.9.9.9, 123.123.123.123',
             ),
             {'detail': 'Incorrect authentication credentials.'},
@@ -77,7 +63,7 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the only IP address in X-Forwarded-For is whitelisted
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender().request_header,
+                HTTP_AUTHORIZATION=hawk_auth_sender(_url()).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4',
             ),
             {'detail': 'Incorrect authentication credentials.'},
@@ -86,7 +72,7 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the only IP address in X-Forwarded-For isn't whitelisted
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender().request_header,
+                HTTP_AUTHORIZATION=hawk_auth_sender(_url()).request_header,
                 HTTP_X_FORWARDED_FOR='123.123.123.123',
             ),
             {'detail': 'Incorrect authentication credentials.'},
@@ -95,7 +81,7 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If third-to-last IP in X-Forwarded-For header is whitelisted
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender().request_header,
+                HTTP_AUTHORIZATION=hawk_auth_sender(_url()).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 124.124.124, 123.123.123.123',
             ),
             {'detail': 'Incorrect authentication credentials.'},
@@ -104,7 +90,7 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If last of 3 IPs in X-Forwarded-For header is whitelisted
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender().request_header,
+                HTTP_AUTHORIZATION=hawk_auth_sender(_url()).request_header,
                 HTTP_X_FORWARDED_FOR='124.124.124, 123.123.123.123, 1.2.3.4',
             ),
             {'detail': 'Incorrect authentication credentials.'},
@@ -121,7 +107,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the Authorization header generated from an incorrect ID
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url(),
                     key_id='incorrect',
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
@@ -132,7 +119,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the Authorization header generated from an incorrect secret
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url(),
                     secret_key='incorrect'
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
@@ -143,8 +131,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the Authorization header generated from an incorrect domain
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
-                    url=_url_incorrect_domain,
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url_incorrect_domain(),
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
             ),
@@ -154,8 +142,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the Authorization header generated from an incorrect path
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
-                    url=_url_incorrect_path,
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url_incorrect_path(),
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
             ),
@@ -165,7 +153,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the Authorization header generated from an incorrect method
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url(),
                     method='POST',
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
@@ -177,7 +166,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # content-type
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url(),
                     content_type='incorrect',
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
@@ -188,7 +178,8 @@ def _auth_sender(key_id='some-id', secret_key='some-secret', url=_url,
             # If the Authorization header generated from incorrect content
             dict(
                 content_type='',
-                HTTP_AUTHORIZATION=_auth_sender(
+                HTTP_AUTHORIZATION=hawk_auth_sender(
+                    _url(),
                     content='incorrect',
                 ).request_header,
                 HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
@@ -218,7 +209,7 @@ def test_if_61_seconds_in_past_401_returned(api_client):
     """
     past = datetime.datetime.now() - datetime.timedelta(seconds=61)
     with freeze_time(past):
-        auth = _auth_sender().request_header
+        auth = hawk_auth_sender(_url()).request_header
     response = api_client.get(
         reverse('activity-stream'),
         content_type='',
@@ -235,7 +226,7 @@ def test_if_61_seconds_in_past_401_returned(api_client):
 @pytest.mark.django_db
 def test_if_authentication_reused_401_returned(api_client):
     """If the Authorization header is reused, then a 401 is returned"""
-    auth = _auth_sender().request_header
+    auth = hawk_auth_sender(_url()).request_header
 
     response_1 = api_client.get(
         _url(),
@@ -262,7 +253,7 @@ def test_empty_object_returned_with_authentication_3_ips(api_client):
     with an extra IP address prepended to the X-Forwarded-For then
     the correct, and authentic, data is returned
     """
-    sender = _auth_sender()
+    sender = hawk_auth_sender(_url())
     response = api_client.get(
         _url(),
         content_type='',
@@ -280,7 +271,7 @@ def test_empty_object_returned_with_authentication(api_client):
     """If the Authorization and X-Forwarded-For headers are correct, then
     the correct, and authentic, data is returned
     """
-    sender = _auth_sender()
+    sender = hawk_auth_sender(_url())
     response = api_client.get(
         _url(),
         content_type='',
